// src/app/api-cinekeep/model/pageResponseFavoriteItemResponse.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FavoriteItemResponse } from './favoriteItemResponse';


export interface PageResponseFavoriteItemResponse { 
    content?: Array<FavoriteItemResponse>;
    page?: number;
    size?: number;
    totalElements?: number;
    totalPages?: number;
}